3.75-alt1
- 3.75
- Updated autoheader.patch
- Updated BuildRequires

+++ nmap/timing.cc
@@ -99,7 +99,7 @@
  *                                                                         *
  ***************************************************************************/
 
-/* $Id: timing.cc,v 1.13 2004/08/31 03:46:19 fyodor Exp $ */
+/* $Id: timing.cc,v 1.14 2004/10/18 16:59:37 fyodor Exp $ */
 
 #include "timing.h"
 #include "NmapOps.h"
@@ -186,7 +186,7 @@ void adjust_timeouts2(const struct timeval *sent,
 		    to->timeout);
 
   if (o.scan_delay)
-    to->timeout = MAX(to->timeout, o.scan_delay * 1000);
+    to->timeout = MAX((unsigned) to->timeout, o.scan_delay * 1000);
 
   if (o.debugging > 3) {
     log_write(LOG_STDOUT, "delta %ld ==> srtt: %d rttvar: %d to: %d\n", delta, to->srtt, to->rttvar, to->timeout);
@@ -222,7 +222,7 @@ void enforce_scan_delay(struct timeval *tv) {
 
   gettimeofday(&now, NULL);
   time_diff = TIMEVAL_MSEC_SUBTRACT(now, lastcall);
-  if (time_diff < o.scan_delay) {  
+  if (time_diff < (int) o.scan_delay) {  
     if (o.debugging > 1) {
       printf("Sleeping for %d milliseconds in enforce_scan_delay()\n", o.scan_delay - time_diff);
     }
